pyproject.toml

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "resume-ranker"
version = "0.1.0"
description = "Smart CV shortlisting with skill matching and semantic similarity"
readme = "README.md"
authors = [{name = "Md Sarfraaz"}]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "rich>=13",
  "pydantic>=2",
  "pandas>=2",
  "numpy>=1.26",
  "scikit-learn>=1.3",
  "rapidfuzz>=3",
  "sentence-transformers>=3",
  "pdfminer.six>=20221105",
  "python-docx>=1",
  "matplotlib>=3.8",
]

[project.optional-dependencies]
cli = ["typer>=0.12", "rich>=13"]
app = ["streamlit>=1.35"]
dev = ["pytest>=8", "pytest-cov>=5", "mypy>=1.10", "ruff>=0.5.0", "black>=24.3.0", "pre-commit>=3.7"]

[project.scripts]
resume-ranker = "resume_ranker.cli:app"

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
addopts = "-q"
